+++ FITple-Frontend/src/App.jsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import reactLogo from "../assets/react.svg";
 import viteLogo from "/vite.svg";
 import "./App.css";
-import Navbar from "../component/Navbar";
+import Navbar from "../component/Navbar/Navbar";
 
 function App() {
   const [count, setCount] = useState(0);
